Le fitBbox se fait plutôt sur la distance, elle toujours calculée
